Address warnings using casts and initialized. Add adjust air timer for player to be based on traveling speed.

+++ 2DTestbed/Code/Controller/AIController.cpp
@@ -45,7 +45,7 @@ namespace
 }
 
 AIController::AIController()
-	: m_timer(CParams::iNumTicks)
+	: m_timer((float)CParams::iNumTicks)
 {
 	m_AnnView = std::make_unique<ANNView>();
